requirements.txt
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
cachetools==5.5.0
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
docker==7.1.0
fastapi==0.115.4
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lightgbm==4.5.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.13.3
numpy==1.26.4
optuna==4.0.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==4.25.5
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
st-theme==1.2.3
starlette==0.41.2
streamlit==1.33.0
streamlit-navigation-bar==3.3.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
